requirements.txt
discord.py==2.3.2
dropbox==12.0.2
piexif==piexif-1.1.3
pillow==10.3.0
ffmpeg==6.1.1